ic_md

Module configs

Structs

ActuatorStatus

Actuator status. This struct is used to keep track of the status of the actuator pins. Upon first initialization they are both set to PinStatus::Low. The actuator pins are ACT0 and ACT1.

CntSetup

Setup for a specific counter. Use this struct to declare the setup of a specific counter.

DeviceStatus

Device Status This struct describes the status of the device. The variables that indicate if a warning or error has occured. This status is updated whenever the counters are read, as errors and warnings are sent along.

FullDeviceStatus

Full Device Status This struct contains the full status of the device that is returned when reading the status registers. For most registers, reading the status will reset the status bits to Ok or the equivalent for the specific status.

Enums

CntCfg

Counter configuration The iC-MD can be configured for 1 up to 3 channels with counter lengths of 16 to 48 bits. Each counter can furthermore be specified to count in clockwise or counterclockwise direction. Finally, you can also configure if the Z signal is normal or inverted. For the setup with three counters, the Z signal setup will simply be ignored as there are no connections for Z signals available. See datasheet for more information.

CntCount

Represent the counter values for different configurations of the iC-MD quadrature counter. If more than one counter value is present, the counter values are always in the order of Counter 0, Counter 1, and Counter 2. Note: The size of the returned value depends on the configuration of the counter!

CntDirection

Enum to specify the direction in which a counter counts This enum is used to turn the positive direction of counting around. By default, it is set to CW for positive counting, but can be set to CCW for positive counting.

CntZSignal

Enum to specify if the Z signal is normal or inverted

CommunicationStatus

Communiucatoion Status This enum indicates if the communication with the device has experienced a collision or not.

DecodificationStatus

Decodification Status A DecodificationError indicates that either the counting frequency is too high or that two incremental edges are too close together.

ErrorStatus

Error Status Enum that indicates if an error has occured or not.

InterfaceStatus

Interface Status This enum indicates if an interface is enabled or disabled.

OverflowStatus

Overflow Status

PinStatus

Status enum for pins. PinStatus::High means that the pin is at VDD, PinStatus::Low means that the pin is at GND.

RegisterStatus

Register Status This enum indicates if a register is valid (Ok) or not (Invalid).

TouchProbeStatus

Touch probe Status This enum indicates if the TPx registers are not loaded / have not been updated or if new values were loaded into the them.

UndervoltageStatus

Power Status If VDD falls below the power off supply level, the device is reset and the RAM initialized to the default value. This status bit indicates that this initialization has taken place (and you might want to consider re-initializing the device).

WarningStatus

Warning Status Enum that indicates if a warning has occured or not.

ZeroStatus

Zero Status This enum indicates if the counter has reached the zero value or not.