ic_md::dd::field_sets

Enum FieldSetValue

pub enum FieldSetValue {

    CounterConfiguration(CounterConfiguration),
    ReadCntCfg0(ReadCntCfg0),
    ReadCntCfg1(ReadCntCfg1),
    ReadCntCfg2(ReadCntCfg2),
    ReadCntCfg3(ReadCntCfg3),
    ReadCntCfg4(ReadCntCfg4),
    ReadCntCfg5(ReadCntCfg5),
    ReadCntCfg6(ReadCntCfg6),
    ReadCntCfg7(ReadCntCfg7),
    ReferenceCounter(ReferenceCounter),
    InstructionByte(InstructionByte),
    Status0(Status0),
    Status1(Status1),
    Status2(Status2),

}

Enum containing all possible field set types

Variants

CounterConfiguration(CounterConfiguration)

Counter configuration The iC-MD can be configured for 1 up to 3 channels with counter lengths of 16 to 48 bits. Here, the counter configuration is selected as a u8 value. The higher-level driver takes care of converting from a meaningful configuration to the 8-bit value.

ReadCntCfg0(ReadCntCfg0)

Read the 24 bit counter configuration, 24+2 bits to read (4 bytes) This corresponds to counter configuration 0b000.

ReadCntCfg1(ReadCntCfg1)

Read the 24 bit, 2 counters configuration, 48+2 bits to read (7 bytes) This corresponds to counter configuration 0b001.

ReadCntCfg2(ReadCntCfg2)

Read the 48 bit counter register, 48+2 bits to read (7 bytes) This corresponds to counter configuration 0b010.

ReadCntCfg3(ReadCntCfg3)

Read the 16 bit counter configuration, 16+2 bits to read (3 bytes) This corresponds to counter configuration 0b011.

ReadCntCfg4(ReadCntCfg4)

Read the 32 bit counter configuration, 32+2 bits to read (5 bytes) This corresponds to counter configuration 0b100.

ReadCntCfg5(ReadCntCfg5)

Read the 32 bit and 16 bit counter configuration, 32+16+2 bits to read (7 bytes) This corresponds to counter configuration 0b101.

ReadCntCfg6(ReadCntCfg6)

Read the 16 bit and 16 bit counter configuration, 16+16+2 bits to read (5 bytes) This corresponds to counter configuration 0b110.

ReadCntCfg7(ReadCntCfg7)

Read the 3 x 16 bit counter configuration, 16+16+16+2 bits to read (7 bytes) This corresponds to counter configuration 0b111.

ReferenceCounter(ReferenceCounter)

Read the references registers 24 bits. TODO: It is unclear if this works, as I assume the address for reading is auto-incremented as when reading the data. This should be tested once the actual hardware setup is available with an encoder connected.

InstructionByte(InstructionByte)

Instruction byte (write only) Allows writing of the instruction bytes. When one of these bits is set to 1, the corresponding instruction is executed and the bit set back to zero, except in the case of Act0 and Act1, which remain set to the written value.

Status0(Status0)

Status0: Status of counter 0 Returns the status of counter 0 plus several other status bits. See also Status1 and Status2 for the other counters and more status bits.

Status1(Status1)

Status1: Status of counter 1 Returns the status of counter 1 plus several other status bits. See also Status0 and Status2 for the other counters and more status bits.

Status2(Status2)

Status2: Status of counter 2 Returns the status of counter 2 plus several other status bits. See also Status0 and Status1 for the other counters and more status bits.