ic_md

Module hl_blocking

This module implements the blocking high-level driver.

This module is available on the “blocking” feature, which is activated by default.

Structs

IcMd

The main driver struct of the crate representing the iC-MD quadrature counter. You can also access the underlying device driver directly via the device field. You are then yourself responsible for reading the correct counter configurations.