Crate ic_md

Driver for the iC-MD quadrature counter. Built fully in Rust, uses embedded_hal and device_driver.

Introduction

The IcMd struct provides a high-level interface to interact with the iC-MD quadrature counter. However, you can also access the underlying device driver directly via the device field. Please read the device driver documentation for more information on what to expect when interfacing with the device driver directly. This low-level access is a temporary solution until the high-level interface is fully developped. When this well be the case is unclear. If you are interested in it, please let me know and I’m happy to prioritize the high-level features that are interesting to you.

Limitations

The following features are currently only accessible via the low-level interface:

Reference register readout: It is unclear if this currently works, see code comment.

The following features are currently not yet implemented:

Input configuration: You can currently not specify the input configuration.

Differential or TTL inputs (Address 0x01, bit 7)

Configuration to have Z signal clear counters 0 and/or 1 (Address 0x01, bits 5 and 6)

Z signal configuration (Address 0x01, bits 3 and 4)

Touch probe and AB registers (Address 0x01, bits 1 and 2)

Differential input configuration selection (RS-422 (default) or LVDS) (Address 0x03, bit 7)

Example Usage

TODO

Further help

TODO, but there should be full tests that help with understanding how to use the driver.

Re-exports

pub use configs::*;

Modules

configs

dd

Structs

IcMd

The main driver struct of the crate representing the iC-MD quadrature counter. You can also access the underlying device driver directly via the device field. You are then yourself responsible for reading the correct counter configurations.