ic_md

Module configs

Module to hold the configuration and status structs for the device

Structs

ActuatorStatus

Actuator status.

CntSetup

Setup for a specific counter.

DeviceStatus

Device Status

FullDeviceStatus

Full Device Status

Enums

CntCfg

Counter configuration

CntCount

Represent the counter values for different configurations of the iC-MD quadrature counter.

CntDirection

Enum to specify the direction in which a counter counts

CntZSignal

Enum to specify if the Z signal is normal or inverted

CommunicationStatus

Communication Status

DecodificationStatus

Decodification Status

ErrorStatus

Error Status

InterfaceStatus

Interface Status This enum indicates if an interface is enabled or disabled.

OverflowStatus

Overflow Status

PinStatus

Status enum for pins.

RegisterStatus

Register Status

TouchProbeStatus

Touch probe Status

UndervoltageStatus

Power Status

WarningStatus

Warning Status

ZeroStatus

Zero Status